+++ CHANGELOG.md
@@ -153,3 +153,16 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### v2.11.5
 - Fix: Error when sending order to fraud analysis and there's not currency code on the order
+
+### v2.12.0
+- Feat: Widget Banner for BNPL
+- Feat: Button Buy with Pix Parcelado with options to show on product page and cart
+- Feat: PCI Compliance option for credit card payments, new config in credit card system configuration
+
+### v2.12.1
+- Fix: Adjust endpoints for PCI Credit Card Form
+
+### v2.13.0:
+- feat: Creating a bnpl modal after payment fail
+- feat: Implement the antifraud strategies
+- fix: installments error when saving data